+++ packages/cubejs-backend-cloud/src/live-preview.ts
@@ -1,6 +1,8 @@
 import chokidar from 'chokidar';
 import { FSWatcher } from 'fs';
 
+import { internalExceptions } from '@cubejs-backend/shared';
+
 import { CubeCloudClient, AuthObject } from './cloud';
 import { DeployController } from './deploy';
 
@@ -15,17 +17,19 @@ export class LivePreviewWatcher {
 
   private queue: {}[] = [];
 
-  public setAuth(token: string) {
+  public setAuth(token: string): AuthObject {
     try {
       const payload = JSON.parse(Buffer.from(token.split('.')[1], 'base64').toString());
       this.auth = {
         auth: token,
-        deploymentId: payload.d,
+        deploymentId: payload.dId,
         deploymentUrl: payload.dUrl,
         url: payload.url,
       };
-    } catch (error) {
-      console.error(error);
+
+      return this.auth;
+    } catch (e) {
+      internalExceptions(e);
       throw new Error('Live-preview token is invalid');
     }
   }
@@ -87,8 +91,8 @@ export class LivePreviewWatcher {
         this.queue = [];
         await this.deploy();
       }
-    } catch (error) {
-      console.error(error);
+    } catch (e) {
+      internalExceptions(e);
     } finally {
       this.handleQueueTimeout = setTimeout(async () => this.handleQueue(), 1000);
     }