+++ packages/cubejs-schema-compiler/src/adapter/BaseQuery.js
@@ -840,19 +840,25 @@ export class BaseQuery {
   }
 
   joinQuery(join, subQueryDimensions) {
+    const subQueryDimensionsByCube = R.groupBy(d => this.cubeEvaluator.cubeNameFromPath(d), subQueryDimensions);
     const joins = join.joins.map(
       j => {
         const [cubeSql, cubeAlias, conditions] = this.rewriteInlineCubeSql(j.originalTo, true);
-        return {
+        return [{
           sql: cubeSql,
           alias: cubeAlias,
           on: `${this.evaluateSql(j.originalFrom, j.join.sql)}${conditions ? ` AND (${conditions})` : ''}`
-        };
+          // TODO handle the case when sub query referenced by a foreign cube on other side of a join
+        }].concat((subQueryDimensionsByCube[j.originalTo] || []).map(d => this.subQueryJoin(d)));
       }
-    ).concat(subQueryDimensions.map(d => this.subQueryJoin(d)));
+    ).reduce((a, b) => a.concat(b), []);
 
     const [cubeSql, cubeAlias] = this.rewriteInlineCubeSql(join.root);
-    return this.joinSql([{ sql: cubeSql, alias: cubeAlias }, ...joins]);
+    return this.joinSql([
+      { sql: cubeSql, alias: cubeAlias },
+      ...(subQueryDimensionsByCube[join.root] || []).map(d => this.subQueryJoin(d)),
+      ...joins
+    ]);
   }
 
   joinSql(toJoin) {
@@ -1122,7 +1128,14 @@ export class BaseQuery {
       .concat(this.segments)
       .concat(this.filters)
       .concat(this.measureFilters)
-      .concat(excludeTimeDimensions ? [] : this.timeDimensions);
+      .concat(excludeTimeDimensions ? [] : this.timeDimensions)
+      .concat(this.join ? this.join.joins.map(j => ({
+        getMembers: () => [{
+          path: () => null,
+          cube: () => this.cubeEvaluator.cubeFromPath(j.originalFrom),
+          definition: () => j.join,
+        }]
+      })) : []);
     return this.collectFrom(membersToCollectFrom, fn, methodName);
   }